pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyeidors"
version = "0.1.0"
description = "PyEidors - A FEniCS-based Framework for Forward and Inverse Problems in Electrical Impedance Tomography (EIT)"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "BingZhou Chen", email = "b20223080651@cau.edu.cn"},
]
keywords = ["EIT", "EIDORS", "FEniCS", "inverse problems", "medical imaging", "pytorch"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Plants/Animals Science Apps",
]

dependencies = [
    "numpy~=1.26.4",      # 兼容 1.26.x 系列
    "scipy~=1.12.0",       # 兼容 1.12.x 系列
    "matplotlib~=3.5.0",   # 兼容 3.5.x 系列
    "pandas~=2.3.0",
    "h5py~=3.11.0",
    "pyyaml~=5.4.1",       # FEniCS 2019.2.0.dev0 依赖 PyYAML 5.4.1
]

[project.optional-dependencies]
torch = [
    "torch~=2.7.1",        # PyTorch 及其相关组件
    "torchvision~=0.22.1", 
    "torchaudio~=2.7.1",
]
cuqi = [
    "CUQIpy~=1.3.0",
    "CUQIpy-FEniCS~=0.8.0",
]
# 开发依赖
dev = [
    "pytest>=6.2.0", 
    "pytest-cov>=3.0.0",
    "black>=22.0.0", 
    "flake8>=4.0.0",
    "pre-commit>=2.15.0"
]
# 文档依赖
docs = [
    "sphinx>=4.0.0", 
    "sphinx-rtd-theme>=1.0.0",
    "jupyter>=1.0.0",
    "nbsphinx>=0.8.0"
]
# 可视化增强
viz = [
    "mayavi>=4.7.0",
    "pyvista>=0.37.0",
    "vtk>=9.0.0"
]

[project.urls]
Homepage = " "
Documentation = " "
Repository = " "
"Bug Tracker" = " "

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pyeidors = ["*.yaml", "*.yml", "data/*"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/pyeidors --cov-report=html --cov-report=term-missing"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests that require GPU",
    "fenics: marks tests that require FEniCS",
    "cuqi: marks tests that require CUQIpy",
]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "docs/source/conf.py", "old", "build", "dist"]